requirements.txt
# LLM Services
langchain==0.3.15
langchain-groq==0.2.1
langchain-core==0.3.31
langchain-community==0.3.15
langgraph==0.2.61

# Web Framework & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5

# Frontend
streamlit==1.41.1
plotly==5.24.1

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.2

# HTTP Client
requests==2.32.3

# Utilities
typing-extensions==4.12.2

# Ontology Processing
rdflib==7.1.1

# Logging & Console
rich==13.9.4

# Embeddings & NLP
cohere==5.12.0
mistralai==1.2.6
numpy==1.26.4
